refactor: centralise ability event handling in AbilityEvents
